Clearlinux Dockerfile: force update of the image
This forces the image to be updated. I previously thought that this
would happen automagically when installing additional bundles but
I was either wrong (or something else escapes me).

The '-b' option is required else I get a fatal (cryptic) error about
running Out Of Memory and a kernel not found.

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>