Clear Linux: add new, optimized Dockerfile
Add (temporarily) a new Dockerfile for Clear Linux that is more
optimized in size.

Note: it does not build the EFI stub as it's missing the efi.h hearder
file. I can be added by also installing the 'os-core-dev' bundle but
that's adding 1GB just for the gnu-efi development package. A request to
the Clear Linux team has been made to create a new 'devpkg-gnu-efi' so
it can be installed in this Clear Linux image.

Signed-off-by: Geoffroy Van Cutsem <geoffroy.vancutsem@intel.com>